backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
requests
groq
GitPython
httpx
pydantic-settings
arq
redis
jose[ cryptography]
passlib[bcrypt]
python-multipart
